requirements.txt
h5py
matplotlib
numpy
Pillow
PyYAML
scikit_learn
scipy
tqdm
tensorflow_gpu==2.9.2
